Chore: Rename users-lists.tsx to index.tsx
